Reorganized and segmented the process of GiantBombAPI. Also made it static, for the moment.
The GetThumb() method now runs through every process needed to check for a game ID of 999999 or default thumbnail and update either, starting with the ID.

A new bool, gameDataChanged, is set to true when any part of the passed game parameter is changed. The last step of GetThumb() is to update the JSON file if this bool is true.